refactor(cashier): consolidate closeCashierSession+cashierHandover into command-consumer-utils
The housekeeping-service had its own SQL implementation of closeCashierSession
(a simple UPDATE) and cashierHandover, duplicating the billing-service versions.
Worse, housekeeping's close skipped payment aggregation, creating a reconciliation
bug: shifts closed from housekeeping would record no transaction totals.

Moved closeCashierSession and cashierHandover into @tartware/command-consumer-utils/cashier
using the same deps-injection pattern as openCashierSession. Both billing-service
and housekeeping-service are now thin parse-then-delegate wrappers.

Removed unused BillingCommandError/SYSTEM_ACTOR_ID/asUuid/resolveActorId from
housekeeping-service's cashier-common.ts (only CommandContext re-export remains).
